+++ CHANGES.md
@@ -1,5 +1,7 @@
 #### 0.10.0
+* use jQuery to remove plugin update row not CSS to hide it
 * use transients not wp_options table for cache as that's what would be used in core
+
 #### 0.9.0
 * add admin notice in `update-core.php`